Pull Changes (#6)
* Update qacheck.yml

* Update qa-on-pull-cert.yml

* Update qa-on-pull-folder.yml

* Update qa-on-push.yml

* Names changed

* workflow modifications

* Workflow renamed

* Update README.md

* transitive trust test added

* update checks

* update workflow

* workflow modified

* workflow modified

* test

* git pull added

* pull modification

* pull removed

* git test

* git update

* fix

* logging added

* update

* test

* test

* test

* fix in script

* test

* test

* test

* clean up

* script fix

* Script fix.

* log disabled

* fix in script

* script fix

* Bot added Files from NLD

* test

* Delete Report

* Delete UP_SYNC.CSR

* script addition

* Create UP_SYNC.PEM

* script fixed

* Delete NLD/onboarding/DCC directory

* Update qa-on-pull-folder.yml

* Delete fileCheck.sh

* Delete review.sh

* Review Bot Test

* workflow modified

* Log added

* skript fix

* script modified

* review bot script modified

* modification

* update

* update

* Workflows renamed

* test

* added value

* fix in script

* Version changed

* Update _test_transitive_trust.py

* Fix in conftest

* Update qa-on-pull-folder.yml

* fix(onboarding): remove participants from DEV

* test

* modify

* modify

* experiment

* test

* fix

* PR extraction added

* test

* test

* test

* test

* test

* test

* modified

* fix

* fix in review

* review script enhanced

* fix in script

* workflow modified

* fix in script

* log added

* test

* test

* review modified

* script fixed

* workflow renamed

* domain fix

* test

* test results activated

* test

* test

* test

* test

* check removed

* test

* check run check added

---------

+++ scripts/delivery.py
@@ -36,64 +36,53 @@ if __name__=='__main__':
                                      flag='😄', name='Test XX', numeric='9990', official_name='Test Country XX' )
     countries = list(pycountry.countries)
 
+    branches = os.popen("git ls-remote --heads").read()
+    print ("branches found:")
+    print(branches)
+
     for country in countries:        
         cCode = country.alpha_2
         if country.alpha_3 in doc:
             cCode= country.alpha_3
         if cCode in doc: 
             try:
-                os.system("echo Try onboarding for " + country.alpha_3)
-                 
-                ################## Prepare the internal structure
-                os.system("rm -rf repo")
-                os.system("rm -rf temp")
-                os.system("mkdir temp")
-                os.system("echo '"+country.alpha_3 + "\n' > temp/country")
-                os.system("echo '"+doc[cCode]+ "' > temp/base64")
-                if os.system("python scripts/config.py") !=0:
-                    raise Exception("Configuration Error")
+                branchName = country.alpha_3+"/onboardingRequest"
                 
-                if os.path.exists("sync"):  
-                   ###############  Transitive Trust
-                   os.system("./scripts/transitiveTrust.sh "+country.alpha_2)
-                   if os.system("python scripts/onboardingRequest.py ./transit/"+os.environ.get("ENV")+"/countries/"+country.alpha_2) != 0:
-                          raise Exception("Onboarding Request failed.")
-                else:
-                    try:       
-                        if os.system("python scripts/repo.py") != 0:
-                              raise Exception("Repository Cloning failed.")
-                          
-                        os.system("./scripts/verify.sh 1> /dev/null")
+                if not branchName in branches:
+                    os.system("echo Try onboarding for " + country.alpha_3)
+                    
+                    ################## Prepare the internal structure
+                    os.system("rm -rf repo")
+                    os.system("rm -rf temp")
+                    os.system("mkdir temp")
+                    os.system("echo '"+country.alpha_3 + "\n' > temp/country")
+                    os.system("echo '"+doc[cCode]+ "' > temp/base64")
+                    if os.system("python scripts/config.py") !=0:
+                        raise Exception("Configuration Error")
+                    
+                    if os.path.exists("sync"):  
+                    ###############  Transitive Trust
+                        os.system("./scripts/transitiveTrust.sh "+country.alpha_2)
+                        if os.system("python scripts/onboardingRequest.py ./transit/"+os.environ.get("ENV")+"/countries/"+country.alpha_2) != 0: 
+                                raise Exception("Onboarding Request failed.")
+                    else:
+                        try:       
+                            if os.system("python scripts/repo.py") != 0:
+                                raise Exception("Repository Cloning failed.")
+                            
+                            os.system("./scripts/verify.sh 1> /dev/null")
+                            
+                            if os.system("python scripts/onboardingRequest.py repo") != 0:
+                                raise Exception("Onboarding Request failed.")
+                        except Exception as Error:
+                            os.system("echo 'Error occoured for onboarding request " + country.alpha_3 +": "+str(Error)+"'") 
                         
-                        if os.system("python scripts/onboardingRequest.py repo") != 0:
-                          raise Exception("Onboarding Request failed.")
-                    except Exception as Error:
-                        os.system("echo 'Error occoured for onboarding request " + country.alpha_3 +": "+str(Error)+"'") 
+                        ######### Create PR 
+                    os.system("./scripts/createPR.sh "+country.alpha_3)
                 
-            
-                
-                
-                ######### Create PR
-                
-                os.system("./scripts/createPR.sh "+country.alpha_3)
-                
-                #####################
-                if os.path.exists(country.alpha_3+"/onboarding/UP"):
-                 os.system("./scripts/fileCheck.sh "+country.alpha_3+"/onboarding/UP")
-                if os.path.exists(country.alpha_3+"/onboarding/TLS"):
-                 os.system("./scripts/fileCheck.sh "+country.alpha_3+"/onboarding/TLS")
-                if os.path.exists(country.alpha_3+"/onboarding/SCA"):
-                  os.system("./scripts/fileCheck.sh "+country.alpha_3+"/onboarding/SCA")
-                if os.path.exists(country.alpha_3+"/onboarding/ISSUER"):
-                 os.system("./scripts/fileCheck.sh "+country.alpha_3+"/onboarding/ISSUER")
-                   
-                ####### Review section
-                
-                #os.system("./scripts/review.sh "+country.alpha_3)
-            
-                #####################
-            
-                os.system("git checkout main > /dev/null 2>&1")
-                os.system("git reset --hard && git clean -f -d > /dev/null 2>&1")
+                    os.system("git checkout main > /dev/null 2>&1")
+                    os.system("git reset --hard && git clean -f -d > /dev/null 2>&1")
+                else:
+                    os.system("echo Skip "+country.alpha_3 + "Branch already exist merge the branch or delete the branch.")
             except Exception as Error:
                 os.system("echo 'Error occoured for onboarding " + country.alpha_3 +": "+ str(Error)+"'")
